Move user patch to updateService
