+++ src/SelfResearch.UserManagement.API/Features/UserManagement/UserManagementService.cs
@@ -39,28 +39,6 @@ public class UserManagementService : IUserManagementService
     }
 
     /// <inheritdoc/>
-    public async Task<Result<UserDto>> PatchUserAsync(int id, JsonPatchDocument<UserDto> patchingDocument)
-    {
-        if (patchingDocument.Operations.Any(x=> x.path.Contains("State")))
-        {
-            return Result.Fail(new ArgumentError(nameof(patchingDocument), "State field cannot be modified via patching."));
-        }
-        var dbUser = await this._userManagementRepository.GetUserAsync(id);
-        if (dbUser == null)
-        {
-            return Result.Fail(new NotFoundError(id.ToString(), nameof(UserDto)));
-        }
-
-        var userDto = this._mapper.Map<UserDto>(dbUser);
-        patchingDocument.ApplyTo(userDto);
-
-        this._mapper.Map(userDto, dbUser);
-
-        await this._userManagementRepository.UpdateUserAsync();
-        return Result.Ok(userDto);
-    }
-
-    /// <inheritdoc/>
     public async Task<Result<bool>> DeleteUserAsync(int id)
     {
         var dbUser = await this._userManagementRepository.GetUserAsync(id);
